Adding source_terms function

+++ init/process_params.jl
@@ -15,7 +15,7 @@ function process_params()
         qi[ispec] = Zi[ispec] * qe #charge of species in statcoulomb
     end
 
-    nquiet = floor(1.18e-2 / abs(dr))
+    nquiet::Int = floor(1.18e-2 / abs(dr))
     cs = sqrt(2 * g * maximum(temp0) / mean(mi)) #it's an "average" speed of sound
     CFL = abs(dr) / cs
     dtm = coeff_CFL * CFL #may need to be varied to avoid NaN when tm > tm_coll
